requirements.txt
streamlit
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage
oauth2client
pandas
folium
streamlit-folium